chore: remove hooks lib
