+++ apps/compositions/src/examples/tabs-lazy-mounted.tsx
@@ -1,8 +1,24 @@
 "use client"
 
-import { useInterval } from "@chakra-ui/hooks"
-import { Tabs } from "@chakra-ui/react"
-import { useState } from "react"
+import { Tabs, useCallbackRef } from "@chakra-ui/react"
+import { useEffect, useState } from "react"
+
+export function useInterval(callback: () => void, delay: number | null) {
+  const fn = useCallbackRef(callback)
+
+  useEffect(() => {
+    let intervalId: number | null = null
+    const tick = () => fn()
+    if (delay !== null) {
+      intervalId = window.setInterval(tick, delay)
+    }
+    return () => {
+      if (intervalId) {
+        window.clearInterval(intervalId)
+      }
+    }
+  }, [delay, fn])
+}
 
 export const TabsLazyMounted = () => {
   return (
