+++ apps/compositions/src/ui/password-input.tsx
@@ -1,13 +1,20 @@
 "use client"
 
-import { mergeRefs, useControllableState } from "@chakra-ui/hooks"
 import type {
   ButtonProps,
   GroupProps,
   InputProps,
   StackProps,
 } from "@chakra-ui/react"
-import { Box, HStack, IconButton, Input, Stack } from "@chakra-ui/react"
+import {
+  Box,
+  HStack,
+  IconButton,
+  Input,
+  Stack,
+  mergeRefs,
+  useControllableState,
+} from "@chakra-ui/react"
 import { forwardRef, useRef } from "react"
 import { LuEye, LuEyeOff } from "react-icons/lu"
 import { InputGroup } from "./input-group"
