--- packages/hooks/README.md
@@ -1,9 +1,0 @@
-# @chakra-ui/hooks
-
-A collection of React Hooks for Chakra UI components.
-
-## Installation
-
-```sh
-yarn add @chakra-ui/hooks
-```
