--- packages/hooks/package.json
@@ -1,66 +1,0 @@
-{
-  "name": "@chakra-ui/hooks",
-  "version": "3.0.0-next.23",
-  "description": "React hooks for Chakra components",
-  "keywords": [
-    "hooks",
-    "react",
-    "chakra-ui",
-    "utilities"
-  ],
-  "type": "module",
-  "sideEffects": false,
-  "author": "Segun Adebayo <sage@adebayosegun.com>",
-  "homepage": "https://github.com/chakra-ui/chakra-ui#readme",
-  "license": "MIT",
-  "main": "dist/cjs/index.cjs",
-  "module": "dist/esm/index.js",
-  "types": "dist/types/index.d.ts",
-  "files": [
-    "dist"
-  ],
-  "publishConfig": {
-    "access": "public"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/chakra-ui/chakra-ui.git",
-    "directory": "packages/legacy/hooks"
-  },
-  "bugs": {
-    "url": "https://github.com/chakra-ui/chakra-ui/issues"
-  },
-  "dependencies": {
-    "@chakra-ui/utils": "workspace:*",
-    "@zag-js/element-size": "0.64.0",
-    "copy-to-clipboard": "3.3.3",
-    "framesync": "6.1.2"
-  },
-  "peerDependencies": {
-    "react": ">=18"
-  },
-  "devDependencies": {
-    "react": "^18.3.1"
-  },
-  "scripts": {
-    "build:fast": "tsx ../../scripts/build/main.ts",
-    "build": "tsx ../../scripts/build/main.ts --dts",
-    "dev": "pnpm build:fast --watch",
-    "clean": "rimraf dist .turbo",
-    "typecheck": "tsc --noEmit"
-  },
-  "exports": {
-    ".": {
-      "source": "./src/index.ts",
-      "import": {
-        "types": "./dist/types/index.d.ts",
-        "default": "./dist/esm/index.js"
-      },
-      "require": {
-        "types": "./dist/types/index.d.ts",
-        "default": "./dist/cjs/index.cjs"
-      }
-    },
-    "./package.json": "./package.json"
-  }
-}
