--- packages/hooks/src/use-safe-layout-effect.ts
@@ -1,7 +1,0 @@
-"use client"
-
-import { useEffect, useLayoutEffect } from "react"
-
-export const useSafeLayoutEffect = Boolean(globalThis?.document)
-  ? useLayoutEffect
-  : useEffect
