--- packages/hooks/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "include": ["src", "__stories__", "__tests__", "../../vitest.setup.ts"],
-  "compilerOptions": {
-    "customConditions": ["source"]
-  }
-}
