+++ packages/react/package.json
@@ -40,7 +40,6 @@
   "dependencies": {
     "@ark-ui/anatomy": "3.5.0",
     "@ark-ui/react": "3.10.0",
-    "@chakra-ui/hooks": "workspace:*",
     "@chakra-ui/utils": "workspace:*",
     "@emotion/is-prop-valid": "1.2.2",
     "@emotion/serialize": "1.1.4",
