+++ packages/react/src/components/index.ts
@@ -79,5 +79,4 @@ export * from "./timeline"
 export * from "./toast"
 export * from "./tooltip"
 export * from "./typography"
-export * from "./use-breakpoint"
 export * from "./visually-hidden"
