+++ packages/react/src/index.ts
@@ -5,3 +5,4 @@ export { mergeRefs } from "./merge-refs"
 export { defaultConfig, defaultSystem } from "./preset"
 export { defaultBaseConfig } from "./preset-base"
 export * from "./styled-system"
+export * from "./hooks"
