+++ sandbox/next-pages/tsconfig.json
@@ -21,7 +21,6 @@
     "customConditions": ["source"],
     "paths": {
       "@chakra-ui/react": ["../../packages/react/src"],
-      "@chakra-ui/hooks": ["../../packages/hooks/src"],
       "@chakra-ui/utils": ["../../packages/utils/src"]
     }
   },
