+++ vite.config.ts
@@ -7,7 +7,6 @@ export default defineConfig({
     alias: {
       "@chakra-ui/utils": resolve("packages/utils/src"),
       "@chakra-ui/react": resolve("packages/react/src"),
-      "@chakra-ui/hooks": resolve("packages/hooks/src"),
       compositions: resolve("apps/compositions/src"),
     },
   },
