+++ website/tsconfig.json
@@ -19,7 +19,6 @@
     "baseUrl": ".",
     "paths": {
       "@chakra-ui/react": ["../packages/react/src"],
-      "@chakra-ui/hooks": ["../packages/hooks/src"],
       "@chakra-ui/utils": ["../packages/utils/src"],
       "contentlayer/generated": ["./.contentlayer/generated"],
       "components/*": ["src/components/*"],
